fix(billing): BUG-M5-002 invoice NPE + BUG-M5-003 reconciliation mock + carry-over compile fixes
- BUG-M5-002 (InvoiceGenerationServiceTest NPE): mock invoiceRepository.save()
  returned null + skipped @PrePersist. Added @BeforeEach stub that mimics
  JPA save (assigns id + generatedAt). 6/6 PASS.
- BUG-M5-003 (BillingReconciliationServiceTest wrong figures): mockRawEventCost()
  helper ignored its cost arg (always returned List.of()). Now builds a real
  MeteringEvent with costUsdCents. 5/5 PASS.
- billing→audit decouple (migration C4): InvoiceGenerationService publishes
  BillingInvoiceGeneratedEvent instead of injecting AuditLogService.
- Carry-over compile fixes for MVP5 session (MeteringEvent, RoiCalculation,
  EsgMetricRepository, TenantContext, TenantService, KafkaProducerService,
  Flink NgsiLdMessage + FlinkMultiTenantConcurrentIT).

+++ backend/src/test/java/com/uip/backend/billing/service/RoiCalculationServiceTest.java
@@ -259,7 +259,7 @@ class RoiCalculationServiceTest {
         event.setTenantId(TENANT_ID);
         event.setBuildingId(buildingId);
         event.setEventType(type);
-        event.setTokenCount(tokenCount);
+        event.setTokenCount((long) tokenCount);
         event.setRecordedAt(Instant.now());
         return event;
     }
